Opravená kompatibilita pre Windows
