+++ include/QOSG/qtcolorpicker.h
