+++ src/main.cpp
@@ -19,7 +19,7 @@ Q_DECLARE_METATYPE(Qt::MouseButton)
 #include "iostream"
 #include "LuaInterface/LuaInterface.h"
 
-#include "dirent.h"
+//#include "dirent.h"
 
 
 int main(int argc, char *argv[])
@@ -31,22 +31,22 @@ qRegisterMetaType<cv::Mat>("Mat");
 #endif
     qRegisterMetaType<Qt::MouseButton>("MouseButton");
 
-    DIR *dir;
-    struct dirent *ent;
-    if ((dir = opendir (".")) != NULL) {
-      /* print all the files and directories within directory */
-      while ((ent = readdir (dir)) != NULL) {
-        printf ("%s\n", ent->d_name);
-      }
-      closedir (dir);
-    } else {
-      /* could not open directory */
-      perror ("");
-      return EXIT_FAILURE;
-    }
+//    DIR *dir;
+//    struct dirent *ent;
+//    if ((dir = opendir (".")) != NULL) {
+//      /* print all the files and directories within directory */
+//      while ((ent = readdir (dir)) != NULL) {
+//        printf ("%s\n", ent->d_name);
+//      }
+//      closedir (dir);
+//    } else {
+//      /* could not open directory */
+//      perror ("");
+//      return EXIT_FAILURE;
+//    }
 
     QApplication app(argc, argv);
     new Util::Cleaner(&app);
     AppCore::Core::getInstance(&app);
-	Manager::GraphManager::getInstance();
+    Manager::GraphManager::getInstance();
 }
